Moves debounce function to utils/debounce
