+++ src/Annotation/AnnotationTable.jsx
@@ -23,22 +23,7 @@ import {
 import AnnotationToolControl from './AnnotationToolControl';
 import ImportAnnotation from './ImportAnnotation';
 import ExportAnnotation from './ExportAnnotation';
-
-function debounce(func, wait, immediate) {
-  let timeout;
-  function debounced(...args) {
-    const context = this;
-    const later = () => {
-      timeout = null;
-      if (!immediate) func.apply(context, args);
-    };
-    const callNow = immediate && !timeout;
-    clearTimeout(timeout);
-    timeout = setTimeout(later, wait);
-    if (callNow) func.apply(context, args);
-  }
-  return debounced;
-}
+import debounce from '../utils/debounce';
 
 function AnnotationTable(props) {
   const {
