+++ src/utils/debounce.js
@@ -1,0 +1,17 @@
+// adapted from underscore.js (https://underscorejs.org/)
+
+export default (func, wait, immediate) => {
+  let timeout;
+  function debounced(...args) {
+    const context = this;
+    const later = () => {
+      timeout = null;
+      if (!immediate) func.apply(context, args);
+    };
+    const callNow = immediate && !timeout;
+    clearTimeout(timeout);
+    timeout = setTimeout(later, wait);
+    if (callNow) func.apply(context, args);
+  }
+  return debounced;
+};
